share some more code with reads
